Move GameState to serializable package
Otherwise, it would crash with the future `Cell`, `Player`, etc.
classes.
