+++ game/src/main/scala/io/aigar/game/Game.scala
@@ -15,16 +15,16 @@ class Game(val id: Int) {
 
   def state = {
     //TODO really implement
-    GameState(
+    io.aigar.game.serializable.GameState(
         id,
         5,
         List(
-          Player(12, "such", 555, List(Cell(5, 5, Position(10,10), Position(10, 10)))),
-          Player(13, "wow", 555, List[Cell]())
+          io.aigar.game.serializable.Player(12, "such", 555, List(io.aigar.game.serializable.Cell(5, 5, io.aigar.game.serializable.Position(10,10), io.aigar.game.serializable.Position(10, 10)))),
+          io.aigar.game.serializable.Player(13, "wow", 555, List[io.aigar.game.serializable.Cell]())
         ),
-        Food(List(Position(5,5)), List[Position](), List[Position]()),
-        Dimensions(10, 10),
-        List[Position]()
+        io.aigar.game.serializable.Food(List(io.aigar.game.serializable.Position(5,5)), List[io.aigar.game.serializable.Position](), List[io.aigar.game.serializable.Position]()),
+        io.aigar.game.serializable.Dimensions(10, 10),
+        List[io.aigar.game.serializable.Position]()
       )
   }
 }
